flow: split string-regexp into own node module.
