+++ src/lib/flow/Kconfig
@@ -87,6 +87,7 @@ source "src/modules/flow/power-supply/Kconfig"
 source "src/modules/flow/process/Kconfig"
 source "src/modules/flow/pwm/Kconfig"
 source "src/modules/flow/random/Kconfig"
+source "src/modules/flow/regexp/Kconfig"
 source "src/modules/flow/robotics/Kconfig"
 source "src/modules/flow/servo-motor/Kconfig"
 source "src/modules/flow/si114x/Kconfig"
